start, pure code movement breaking up and consolidating vars.rs
